ABL 64-bit code fixes

+++ code/ablmc2.cpp
@@ -4732,7 +4732,7 @@ void execSendMessage (void) {
 
 void execGetMessage (void) {
 
-	long* messageParam = ABLi_peekIntegerPtr();
+	int* messageParam = ABLi_peekIntegerPtr();
 
 	*messageParam = CurMultiplayParam;
 	ABLi_pokeInteger(CurMultiplayCode);
